pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "researchclaw"
version = "0.3.1"
description = "AutoResearchClaw — Autonomous Research Pipeline"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "AutoResearchClaw Team" }]
keywords = ["research", "automation", "llm", "pipeline", "academic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "PyYAML>=6.0",
    "rich>=13.0",
    "arxiv>=2.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
anthropic = ["httpx>=0.25"]
web = ["scholarly>=1.7", "requests>=2.31", "beautifulsoup4>=4.12"]
pdf = ["PyPDF2>=3.0"]
all = [
    "researchclaw[anthropic,web,pdf]",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "matplotlib>=3.8",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[project.scripts]
researchclaw = "researchclaw.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["researchclaw"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
